+++ src/lib/output/plugins/index.ts
@@ -1,5 +1,5 @@
 export { AssetsPlugin } from "./AssetsPlugin";
 export { JavascriptIndexPlugin } from "./JavascriptIndexPlugin";
 export { MarkedLinksPlugin } from "./MarkedLinksPlugin";
-export { MarkedPlugin as MarkedPlugin } from "../themes/MarkedPlugin";
+export { MarkedPlugin } from "../themes/MarkedPlugin";
 export { LegendPlugin } from "./LegendPlugin";