+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ### Features
 
 -   Flag option types like `validation` can now be set to true/false to enable/disable all flags within them.
+-   Added `githubPages` option (default: true), which will create a `.nojekyll` page in the generated output.
+
+### Thanks!
+
+-   @srmagura
 
 ## v0.22.3 (2021-08-12)